make OtherTestValidationNotice a private static class
