+++ core/src/test/java/org/mobilitydata/gtfsvalidator/notice/NoticeTest.java
@@ -3,6 +3,7 @@ package org.mobilitydata.gtfsvalidator.notice;
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.common.collect.ImmutableMap;
+import java.util.Map;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
@@ -42,5 +43,21 @@ public class NoticeTest {
         .isNotEqualTo(
             new TestValidationNotice(
                 "code1", ImmutableMap.of("key1", "value1"), SeverityLevel.INFO));
+  }
+
+  private static class OtherTestValidationNotice extends ValidationNotice {
+
+    private final String code;
+
+    public OtherTestValidationNotice(
+        String code, Map<String, Object> context, SeverityLevel severityLevel) {
+      super(context, severityLevel);
+      this.code = code;
+    }
+
+    @Override
+    public String getCode() {
+      return code;
+    }
   }
 }
